delete the route that removes a customer with its logic
Customer's email is in the booking table as a foreign key. We can either replace it with the NULL value or, delete all bookings associated with the delete customer. Both of them are not ideal. So, for now, we'll not allow to delete an existing user until the database design is improved to prserve booking inforamation whenever a customer is deleted from the customer table.